Refactor and create a new folder structure. Small hack needed for num_layers to be passed in DS Attention due to this.
