+++ deepspeed/model_implementations/transformers/clip_encoder.py
