+++ deepspeed/ops/transformer/__init__.py
@@ -1,4 +1,5 @@
 from .transformer import DeepSpeedTransformerLayer, DeepSpeedTransformerConfig
-from .inference.transformer_inference import DeepSpeedTransformerInference, DeepSpeedInferenceConfig
+from .inference.config import DeepSpeedInferenceConfig
+from ...model_implementations.transformers.ds_transformer import DeepSpeedTransformerInference
 from .inference.moe_inference import DeepSpeedMoEInferenceConfig, DeepSpeedMoEInference
 from .inference.attention import DeepSpeedAttention
