+++ deepspeed/ops/transformer/inference/__init__.py
@@ -1,3 +1,4 @@
-from .transformer_inference import DeepSpeedTransformerInference, DeepSpeedInferenceConfig
+from .config import DeepSpeedInferenceConfig
+from ....model_implementations.transformers.ds_transformer import DeepSpeedTransformerInference
 from .moe_inference import DeepSpeedMoEInferenceConfig, DeepSpeedMoEInference
 from .attention import DeepSpeedAttention
